v0.0 Implementación multilenguaje 100%. Todos las funcionalidades al 100%. Falta procesos en envíos de mail y 10% de diseño gráfico.

+++ src/homeSweetHome/controller/budget/UpdateSpentViewController.java
@@ -1,19 +1,19 @@
 package homeSweetHome.controller.budget;
 
 import homeSweetHome.dataPersistence.BudgetDAO;
-import homeSweetHome.dataPersistence.CurrentSession;
 import homeSweetHome.model.Budget;
 import homeSweetHome.utils.AlertUtils;
 import homeSweetHome.utils.LanguageManager;
-import java.net.URL;
-import java.util.ResourceBundle;
+import java.util.Optional;
 import javafx.application.Platform;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
+import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 
@@ -37,16 +37,16 @@ public class UpdateSpentViewController {
     private Button btnDeleteSpent;
     @FXML
     private Button btnCancelSpent;
+    @FXML
+    private Label modifySpentTitle;
 
     private Budget budget; // El gasto a modificar
 
     private BudgetViewController budgetViewController; // Referencia al controlador principal
 
-
     /**
      * Método para inicializar el controlador.
      */
-    @FXML
     public void initialize() {
 
 /////////////////////////////////IDIOMAS/////////////////////////////////////////////
@@ -55,7 +55,7 @@ public class UpdateSpentViewController {
         LanguageManager.getInstance().addListener(() -> Platform.runLater(this::updateTexts));
         updateTexts(); // Actualiza los textos inicialmente
 
-/////////////////////////////////FIN IDIOMAS/////////////////////////////////////////////     
+/////////////////////////////////FIN IDIOMAS/////////////////////////////////////////////  
 
     }
 
@@ -65,17 +65,17 @@ public class UpdateSpentViewController {
      * Actualiza los textos de la interfaz en función del idioma.
      */
     private void updateTexts() {
-        
+
         /// Accede directamente al Singleton del LanguageManager
         LanguageManager languageManager = LanguageManager.getInstance();
 
         if (languageManager == null) {
-            
+
             System.err.println("Error: LanguageManager no está disponible.");
             return;
-            
+
         } else {
-            
+
             System.out.println("Método updateTexts llamado en UpdateSpentView. Idioma actual: " + languageManager.getLanguageCode());
         }
 
@@ -87,6 +87,8 @@ public class UpdateSpentViewController {
         System.out.println("Texto de btnUpdateSpent: " + btnUpdateSpent.getText());
         System.out.println("Texto de btnCancelSpent: " + btnCancelSpent.getText());
         System.out.println("Texto de btnDeleteSpent: " + btnDeleteSpent.getText());
+
+        modifySpentTitle.setText(languageManager.getTranslation("modifySpentTitle"));
 
         // Traducción de los campos de texto con PromptText
         System.out.println("Configurando PromptText de TextField...");
@@ -111,6 +113,7 @@ public class UpdateSpentViewController {
                 languageManager.getTranslation("categoryHousing"),
                 languageManager.getTranslation("categoryOthers")
         );
+        
         System.out.println("Elementos actuales de Category: " + Category.getItems());
 
         // Configuración dinámica del ComboBox: Métodos de Pago
@@ -123,6 +126,7 @@ public class UpdateSpentViewController {
                 languageManager.getTranslation("paymentCash"),
                 languageManager.getTranslation("paymentCard")
         );
+        
         System.out.println("Elementos actuales de PaymentMethod: " + PaymentMethod.getItems());
 
         System.out.println("Traducciones aplicadas correctamente en UpdateSpentView.");
@@ -137,7 +141,7 @@ public class UpdateSpentViewController {
      * @param budget El gasto seleccionado en la tabla.
      */
     public void setBudget(Budget budget) {
-        
+
         this.budget = budget;
         loadSpentData();
     }
@@ -170,112 +174,112 @@ public class UpdateSpentViewController {
      */
     @FXML
     private void updateSpent() {
-        
+
         try {
+            
             // Usa el Singleton directamente para traducciones
             LanguageManager languageManager = LanguageManager.getInstance();
 
             if (languageManager == null) {
-                
+
                 System.err.println("Error: LanguageManager no está disponible.");
                 AlertUtils.showAlert(Alert.AlertType.ERROR, "Error", "Gestión de idiomas no disponible.");
                 return;
             }
 
-            // Validar el método de pago seleccionado
+            // Valida el método de pago seleccionado
             String metodoPagoSeleccionado = PaymentMethod.getSelectionModel().getSelectedItem();
-            
+
             if (metodoPagoSeleccionado == null || metodoPagoSeleccionado.trim().isEmpty()) {
-                
+
                 AlertUtils.showAlert(
                         Alert.AlertType.WARNING,
                         languageManager.getTranslation("invalidInput"),
                         languageManager.getTranslation("paymentMethodNotSelected")
                 );
-                
+
                 return;
             }
 
-            // Validar la categoría seleccionada
+            // Valida la categoría seleccionada
             String categoriaSeleccionada = Category.getSelectionModel().getSelectedItem();
-            
+
             if (categoriaSeleccionada == null || categoriaSeleccionada.trim().isEmpty()) {
-                
+
                 AlertUtils.showAlert(
                         Alert.AlertType.WARNING,
                         languageManager.getTranslation("invalidInput"),
                         languageManager.getTranslation("categoryNotSelected")
                 );
-                
+
                 return;
             }
 
-            // Validar los datos de los campos de texto
+            // Valida los datos de los campos de texto
             String nombreGasto = fieldSpentName.getText().trim();
-            
+
             if (nombreGasto.isEmpty()) {
-                
+
                 AlertUtils.showAlert(
                         Alert.AlertType.WARNING,
                         languageManager.getTranslation("invalidInput"),
                         languageManager.getTranslation("spentNameEmpty")
                 );
-                
+
                 return;
             }
 
             String cantidadGasto = fieldSpentQuantity.getText().trim();
             double montoGasto;
-            
+
             try {
-                
+
                 montoGasto = Double.parseDouble(cantidadGasto);
-                
+
                 if (montoGasto <= 0) {
-                    
+
                     throw new NumberFormatException("Monto inválido.");
-                    
+
                 }
-                
+
             } catch (NumberFormatException ex) {
-                
+
                 AlertUtils.showAlert(
                         Alert.AlertType.WARNING,
                         languageManager.getTranslation("invalidInput"),
                         languageManager.getTranslation("spentQuantityInvalid")
                 );
-                
+
                 return;
             }
 
-            // Actualizar el objeto Budget con los datos validados
+            // Actualiza el objeto Budget con los datos validados
             budget.setNombre(nombreGasto);
             budget.setMonto(montoGasto);
-            budget.setMetodoPago(metodoPagoSeleccionado); // Almacenar en el idioma de la interfaz
-            budget.setCategoria(categoriaSeleccionada); // Almacenar en el idioma de la interfaz
+            budget.setMetodoPago(metodoPagoSeleccionado); // Almacena en el idioma de la interfaz
+            budget.setCategoria(categoriaSeleccionada); // Almacena en el idioma de la interfaz
             budget.setFecha(Date.getValue());
             budget.setDescripcion(fieldSpentDescription.getText().trim());
 
-            // Actualizar en la base de datos
+            // Actualiza en la base de datos
             BudgetDAO budgetDAO = new BudgetDAO();
             boolean success = budgetDAO.updateBudget(budget);
 
-            // Mostrar alerta según el resultado de la actualización
+            // Muestra alerta según el resultado de la actualización
             if (success) {
-                
+
                 AlertUtils.showAlert(
                         Alert.AlertType.INFORMATION,
                         languageManager.getTranslation("success"),
                         languageManager.getTranslation("spentUpdated")
                 );
-                
+
                 budgetViewController.loadSpents(); // Refresca la tabla en el controlador principal
                 ((Stage) btnUpdateSpent.getScene().getWindow()).close(); // Cierra la ventana
-                
+
             } else {
-                
+
                 AlertUtils.showAlert(
-                        
                         Alert.AlertType.ERROR,
                         languageManager.getTranslation("error"),
                         languageManager.getTranslation("spentUpdateFailed")
@@ -283,12 +287,11 @@ public class UpdateSpentViewController {
             }
 
         } catch (Exception e) {
-            
+
             e.printStackTrace();
-            LanguageManager languageManager = LanguageManager.getInstance(); 
-            
+            LanguageManager languageManager = LanguageManager.getInstance();
+
             AlertUtils.showAlert(
-                    
                     Alert.AlertType.ERROR,
                     languageManager.getTranslation("error"),
                     languageManager.getTranslation("unexpectedError")
@@ -302,26 +305,62 @@ public class UpdateSpentViewController {
     @FXML
     private void deleteSpent() {
 
-        try {
+        // Crea alerta de confirmación
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.getDialogPane().getStylesheets().add(getClass().getResource("/homeSweetHome/utils/alertsCss.css").toExternalForm());
 
-            BudgetDAO budgetDAO = new BudgetDAO();
-            boolean success = budgetDAO.deleteBudget(budget.getId());
+        //alert.getDialogPane().getStylesheets().add(getClass().getResource("../homeSweetHome/utils/alertsCss.css").toExternalForm());
+        alert.setTitle(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Confirmación" : "Confirmation");
+        alert.setHeaderText(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Eliminar gasto" : "Delete Expense");
+        alert.setContentText(LanguageManager.getInstance().getLanguageCode().equals("es") ? "¿Seguro que desea eliminar este gasto?" : "Are you sure you want to delete this expense?");
 
-            if (success) {
+        // Agrega botones manualmente
+        alert.getButtonTypes().setAll(ButtonType.OK, ButtonType.CANCEL);
 
-                AlertUtils.showAlert(Alert.AlertType.INFORMATION, "Éxito", "El gasto se ha eliminado correctamente.");
-                budgetViewController.loadSpents(); // Refresca la tabla en el controlador principal
-                ((Stage) btnDeleteSpent.getScene().getWindow()).close(); // Cierra la ventana
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
 
-            } else {
+            try {
+                
+                BudgetDAO budgetDAO = new BudgetDAO();
+                boolean success = budgetDAO.deleteBudget(budget.getId());
 
-                AlertUtils.showAlert(Alert.AlertType.ERROR, "Error", "Hubo un problema al intentar eliminar el gasto.");
+                if (success) {
+                    
+                    Alert successAlert = new Alert(Alert.AlertType.INFORMATION);
+                    successAlert.getDialogPane().getStylesheets().add(getClass().getResource("/homeSweetHome/utils/alertsCss.css").toExternalForm());
+                    successAlert.setTitle(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Éxito" : "Success");
+                    successAlert.setHeaderText(null);
+                    successAlert.setContentText(LanguageManager.getInstance().getLanguageCode().equals("es") ? "El gasto se ha eliminado correctamente." : "The expense has been deleted successfully.");
+                    successAlert.showAndWait();
+
+                    budgetViewController.loadSpents(); // Refresca la tabla en el controlador principal
+                    ((Stage) btnDeleteSpent.getScene().getWindow()).close(); // Cierra la ventana
+
+                } else {
+                    
+                    Alert errorAlert = new Alert(Alert.AlertType.ERROR);
+                    errorAlert.getDialogPane().getStylesheets().add(getClass().getResource("/homeSweetHome/utils/alertsCss.css").toExternalForm());
+                    errorAlert.setTitle("Error");
+                    errorAlert.setHeaderText(null);
+                    errorAlert.setContentText(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Hubo un problema al intentar eliminar el gasto." : "There was a problem deleting the expense.");
+                    errorAlert.showAndWait();
+                }
+
+            } catch (Exception e) {
+                
+                e.printStackTrace();
+                Alert errorAlert = new Alert(Alert.AlertType.ERROR);
+                errorAlert.getDialogPane().getStylesheets().add(getClass().getResource("/homeSweetHome/utils/alertsCss.css").toExternalForm());
+                errorAlert.setTitle("Error");
+                errorAlert.setHeaderText(null);
+                errorAlert.setContentText(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Se produjo un error al intentar eliminar el gasto." : "An error occurred while trying to delete the expense.");
+                errorAlert.showAndWait();
             }
-
-        } catch (Exception e) {
-
-            e.printStackTrace();
-            AlertUtils.showAlert(Alert.AlertType.ERROR, "Error", "Se produjo un error al intentar eliminar el gasto.");
+            
+        } else {
+            
+            System.out.println(LanguageManager.getInstance().getLanguageCode().equals("es") ? "Eliminación cancelada por el usuario." : "Deletion canceled by user.");
         }
     }
 
@@ -332,7 +371,7 @@ public class UpdateSpentViewController {
      */
     @FXML
     private void cancel(ActionEvent event) {
-        
+
         ((Stage) btnCancelSpent.getScene().getWindow()).close(); // Cierra la ventana actual
     }
 }
